+++ src/main/java/com/facts/financial_facts_service/repositories/DiscountRepository.java
@@ -1,5 +1,6 @@
-package com.facts.financial_facts_service.entities.discount;
+package com.facts.financial_facts_service.repositories;
 
+import com.facts.financial_facts_service.entities.discount.Discount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
