+++ src/main/java/com/facts/financial_facts_service/repositories/IdentityRepository.java
@@ -1,5 +1,6 @@
-package com.facts.financial_facts_service.entities.identity;
+package com.facts.financial_facts_service.repositories;
 
+import com.facts.financial_facts_service.entities.identity.Identity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
