+++ src/main/java/com/facts/financial_facts_service/services/FactsService.java
@@ -1,23 +1,20 @@
-package com.facts.financial_facts_service.entities.facts;
+package com.facts.financial_facts_service.services;
 
 import com.facts.financial_facts_service.components.WebClientFactory;
 import com.facts.financial_facts_service.constants.Constants;
-import com.facts.financial_facts_service.constants.ModelType;
-import com.facts.financial_facts_service.entities.identity.Identity;
-import com.facts.financial_facts_service.exceptions.DataNotFoundException;
+import com.facts.financial_facts_service.entities.facts.Facts;
+import com.facts.financial_facts_service.repositories.FactsRepository;
 import jakarta.annotation.PostConstruct;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
 
-import java.util.Map;
 import java.util.Optional;
 
 
@@ -52,7 +49,7 @@ public class FactsService implements Constants {
         logger.info("In facts service retrieving facts for cik {}", cik);
         return factsWebClient.get()
             .uri(uriBuilder -> uriBuilder
-                .queryParam("cik", cik)
+                .queryParam(LOWER_CIK, cik)
                 .build())
             .retrieve()
             .bodyToMono(String.class)
