+++ src/main/resources/application.properties
@@ -9,4 +9,4 @@ spring.jpa.properties.hibernate.format_sql=true
 sec.cik.api.endpoint=https://www.sec.gov/files/company_tickers.json
 sec.cik.api.user-agent=your@gmail.com
 
-facts-gateway.baseUrl=${facts-gateway.baseUrl}
+facts-gateway.baseUrl=https://op2y8wcs73.execute-api.us-east-1.amazonaws.com/get-facts-with-cik
