+++ src/test/java/com/facts/financial_facts_service/entities/discount/DiscountControllerTest.java
@@ -1,6 +1,8 @@
 package com.facts.financial_facts_service.entities.discount;
 
 import com.facts.financial_facts_service.constants.TestConstants;
+import com.facts.financial_facts_service.controllers.DiscountController;
+import com.facts.financial_facts_service.services.DiscountService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
